+++ docs/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2021-05-18
+
+### Added
+
+- Ability to switch between Overdrive and Libby library platforms
+- Default to Overdrive (requires less configuration, no sign-in to search)
+- Onboarding windows for new users and users updating to 2.0.0
+
 ## [1.1.0] - 2021-05-12
 
 ### Added